+++ debug/jt/jt-808-server-spring-boot-starter-reactive-debug/src/main/java/io/github/hylexus/xtream/debug/ext/jt808/handler/DemoLocationMessageHandler.java
@@ -24,7 +24,9 @@ import io.github.hylexus.xtream.codec.ext.jt808.extensions.handler.Jt808RequestH
 import io.github.hylexus.xtream.codec.ext.jt808.extensions.handler.Jt808ResponseBody;
 import io.github.hylexus.xtream.codec.ext.jt808.spec.Jt808ProtocolVersion;
 import io.github.hylexus.xtream.codec.ext.jt808.spec.Jt808Request;
+import io.github.hylexus.xtream.codec.ext.jt808.spec.Jt808RequestEntity;
 import io.github.hylexus.xtream.codec.ext.jt808.spec.Jt808Session;
+import io.github.hylexus.xtream.debug.ext.jt808.message.DemoLocationMsg03;
 import io.github.hylexus.xtream.debug.ext.jt808.service.DemoLocationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -51,11 +53,42 @@ public class DemoLocationMessageHandler {
     public Mono<ServerCommonReplyMessage> processMessage0200V2019(
             Jt808Session session,
             Jt808Request request,
-            @Jt808RequestBody BuiltinMessage0200 body) {
+            @Jt808RequestBody BuiltinMessage0200 body,
+            Jt808RequestEntity<BuiltinMessage0200> requestEntity,
+            DemoLocationMsg03 message) {
         log.info("v2019-0x0200: {}", body);
+        log.info("v2019-0x0200: {}", requestEntity);
+        log.info("v2019-0x0200: {}", message);
         return this.locationService.processLocationMessage(session, body).map(result -> {
             // ...
             return ServerCommonReplyMessage.of(request, result);
         });
     }
+
+    // /**
+    //  * 位置上报(V2019)
+    //  * <p>
+    //  * 7e02004086010000000001893094655200E4000000000000000101D907F2073D336C000000000000211124114808010400000026030200003001153101002504000000001404000000011504000000FA160400000000170200001803000000EA10FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF02020000EF0400000000F31B017118000000000000000000000000000000000000000000000000567e
+    //  */
+    // @Jt808RequestHandlerMapping(messageIds = 0x0200, versions = Jt808ProtocolVersion.VERSION_2019)
+    // public Mono<Jt808ResponseEntity<ServerCommonReplyMessage>> processMessage0200V2019(
+    //         // public Jt808ResponseEntity<ServerCommonReplyMessage> processMessage0200V2019(
+    //         Jt808Session session,
+    //         Jt808Request request,
+    //         @Jt808RequestBody BuiltinMessage0200 body,
+    //         Jt808RequestEntity<BuiltinMessage0200> requestEntity,
+    //         DemoLocationMsg03 message) {
+    //     log.info("v2019-0x0200: {}", body);
+    //     log.info("v2019-0x0200: {}", requestEntity);
+    //     log.info("v2019-0x0200: {}", message);
+    //
+    //     // return Mono.just(responseEntity);
+    //     return this.locationService.processLocationMessage(session, body).map(result -> {
+    //         // ...
+    //         return Jt808ResponseEntity
+    //                 .messageId(0x8001)
+    //                 .maxPackageSize(1000)
+    //                 .body(ServerCommonReplyMessage.of(request, (byte) 0));
+    //     });
+    // }
 }